requirements.txt
Amara
Django
Fabric
PIL
PyYAML
South
celery
celery-with-redis
django-annoying
django-celery
django-debug-toolbar
django-mediagenerator
django-supervisor
django-tastypie
django-pipeline
django-extensions
git+git://github.com/jsocol/django-adminplus#egg=django-adminplus
werkzeug
gunicorn
redis
urllib3
